Gravity?

+++ src/main/WindowElements/BaseElement.java
@@ -1,11 +1,11 @@
 package main.WindowElements;
 
 public class BaseElement {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     public String filename;
 
-    public BaseElement(int x, int y, String filename) {
+    public BaseElement(double x, double y, String filename) {
         this.x = x;
         this.y = y;
         this.filename = filename;
