+++ src/main/WindowElements/CircleElement.java
@@ -1,7 +1,7 @@
 package main.WindowElements;
 
 public class CircleElement extends BaseElement {
-    public CircleElement(int x, int y, String filename) {
+    public CircleElement(double x, double y, String filename) {
         super(x, y, filename);
     }
 }
